Refactoring logic for using filters for base64pem
