+++ tests/scanner/scan_type/test_pem_key_pattern.py
@@ -1,6 +1,6 @@
 import unittest
 
-from credsweeper.scanner.scan_type import PemKeyPattern
+from credsweeper.utils.pem_key_detector import PemKeyDetector
 
 
 class TestPemKeyPattern(unittest.TestCase):
@@ -8,7 +8,7 @@ class TestPemKeyPattern(unittest.TestCase):
     def test_remove_leading_config_lines_p(self):
         lines = ["Proc-Type: 4,ENCRYPTED", "DEK-Info: DES-EDE3-CBC,BA2D3F11273F6I7A", ""]
         for line in lines:
-            self.assertTrue(PemKeyPattern.is_leading_config_line(line), line)
+            self.assertTrue(PemKeyDetector.is_leading_config_line(line), line)
 
     def test_remove_leading_config_lines_n(self):
         lines = [
@@ -17,7 +17,7 @@ class TestPemKeyPattern(unittest.TestCase):
             "1s1SPrCBkedbNf0Tp0GbMJDyR4e9T04ZZwIDAQABAoGAFijko56+qGyN8M0RVyaRAXz++xTqHBLh"
         ]
         for line in lines:
-            self.assertFalse(PemKeyPattern.is_leading_config_line(line), line)
+            self.assertFalse(PemKeyDetector.is_leading_config_line(line), line)
 
     def test_sanitize_line_p(self):
         lines = [
@@ -31,7 +31,7 @@ class TestPemKeyPattern(unittest.TestCase):
             "//tDzrYypSREYpBHTUKoa+y0rRy74nLA1Z4+nKVOTdXNuMGLp9KxHqwIlDk8QK5n"
         ]
         for expect_line, line in zip(should_be, lines):
-            self.assertEqual(expect_line, PemKeyPattern.sanitize_line(line), line)
+            self.assertEqual(expect_line, PemKeyDetector.sanitize_line(line), line)
 
     def test_sanitize_line_n(self):
         """Check that valid PEM lines will not be changed"""
@@ -40,4 +40,4 @@ class TestPemKeyPattern(unittest.TestCase):
             "hNtDzrYypSREYpBHTUKoa+y0rRy74nLA1Z4+nKVOTdXNuMGLp9KxHqwIlDk8QK5n"
         ]
         for line in lines:
-            self.assertEqual(line, PemKeyPattern.sanitize_line(line), line)
+            self.assertEqual(line, PemKeyDetector.sanitize_line(line), line)
